[CUMULUS-2020] Fix queries to use the right field type when needed
